Put classes in separate files
