+++ eng/update-dependencies/ChiselUpdater.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.DotNet.VersionTools.Dependencies;
@@ -43,24 +41,4 @@ internal static class ChiselUpdater
     private static Regex GetAssetRegex(string arch) => new(@"chisel_v\d+\.\d+\.\d+_linux_" + arch + @"\.tar\.gz");
 
     private static string ToManifestArch(string arch) => arch == "amd64" ? "x64" : arch;
-
-    /// <summary>
-    /// Updates the checksum when runtime dependencies are being updated.
-    /// </summary>
-    private class ChiselReleaseShaUpdater(string repoRoot, string variableName, Release release, string arch)
-        : GitHubReleaseUrlUpdater(repoRoot, variableName, release, DependencyInfoToUse, GetAssetRegex(arch))
-    {
-        private static readonly HttpClient s_httpClient = new();
-
-        protected override string? GetValue()
-        {
-            string? downloadUrl = base.GetValue();
-            if (downloadUrl is null)
-            {
-                return null;
-            }
-
-            return ChecksumHelper.ComputeChecksumShaAsync(s_httpClient, downloadUrl).Result;
-        }
-    }
 }
